% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metadata.R
\name{check_dataset_model}
\alias{check_dataset_model}
\title{Check dataset model}
\usage{
check_dataset_model(dataset)
}
\arguments{
\item{dataset}{A `data.frame` or `data.table` representing the input dataset.}
}
\value{
Invisibly returns `NULL` on success, otherwise throws an error.
}
\description{
Ensure `dataset` contains the minimal variables required by the
cleaning pipeline (`concept_id`, `value`, `unit`).
}
\keyword{internal}
