% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{CleanLabValues-package}
\alias{CleanLabValues}
\alias{CleanLabValues-package}
\title{CleanLabValues package}
\description{
Ingests instructions to clean datasets containing results from laboratory analyses. Handles unit conversions, missing-unit imputation, and exclusion of physiologically implausible values based on configurable metadata (target units, conversion rules, and thresholds).
}
\author{
\strong{Maintainer}: Yinan Mao \email{ymao@example.com}

Authors:
\itemize{
  \item Yinan Mao \email{ymao@example.com}
  \item Rosa Gini
}

}
\keyword{internal}
