% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanLabValuesDataset.R
\name{CleanLabValuesDataset}
\alias{CleanLabValuesDataset}
\title{CleanLabValuesDataset}
\usage{
CleanLabValuesDataset(
  dataset,
  list_analyses = c(),
  lab_target_units,
  lab_unit_conversion,
  lab_thresholds,
  datasource = ""
)
}
\arguments{
\item{dataset}{the name of a data.table file in memory that contains a dataset of results of laboratory analyses that needs cleaning}

\item{list_analyses}{a string vector containing the names of the laboratory analyses to be cleaned. If the argument is not specified, all the laboratory analyses are cleaned}

\item{lab_target_units}{a string containing the path towards a csv files containing one record per each type of laboratory analysis in list_analyses and specifying the desired unit of measurement}

\item{lab_unit_conversion}{a string containing the path towards a csv files containing the specifications to convert the values in the dataset to the target unit of measurement}

\item{lab_thresholds}{a string containing the path towards a csv files containing the specifications of which values should be considered absurd and discarded, possibly depending on other variables such as age}

\item{datasource}{(non mandatory) a string containing name of the datasource that can be stored in `lab_unit_conversion` to produce a datasource-specific assumption on what to do if the unit of measurement is missing}
}
\value{
A `data.table` with the cleaned lab results and appended result columns.
}
\description{
The function CleanLabValuesDataset ingests instructions to clean datasets containing results from laboratory analysis. The instructions specify which unit of measurement is desired for each laboratory analysis, what the conversion rules are, what to do if unit of measurement is missing, what the values should be considered absurd and discarded.
}
\details{
...
}
\seealso{
...
}
